Move DataFormField and DataFormFieldProps types to FormFields
