+++ mathesar_ui/src/systems/data-forms/config-panel/field-config/FieldAppearance.svelte
@@ -8,7 +8,7 @@
     isDefinedNonNullable,
   } from '@mathesar-component-library';
 
-  import type { DataFormField } from '../../data-form-utilities/types';
+  import type { DataFormField } from '../../data-form-utilities/FormFields';
 
   export let field: DataFormField;
