+++ mathesar_ui/src/systems/data-forms/elements/DataFormLabel.svelte
@@ -10,7 +10,7 @@
     type DataFormManager,
     EditableDataFormManager,
   } from '../data-form-utilities/DataFormManager';
-  import type { DataFormField } from '../data-form-utilities/types';
+  import type { DataFormField } from '../data-form-utilities/FormFields';
 
   import FormFieldCommonControls from './FormFieldCommonControls.svelte';
