+++ mathesar_ui/src/systems/data-forms/elements/FormFieldCommonControls.svelte
@@ -2,7 +2,7 @@
   import { iconDeleteMajor } from '@mathesar/icons';
   import { Button, Icon } from '@mathesar-component-library';
 
-  import type { DataFormField } from '../data-form-utilities/types';
+  import type { DataFormField } from '../data-form-utilities/FormFields';
 
   export let dataFormField: DataFormField;
 </script>
