+++ mathesar_ui/src/systems/data-forms/elements/FormFieldElementWrapper.svelte
@@ -3,7 +3,7 @@
     type DataFormManager,
     EditableDataFormManager,
   } from '../data-form-utilities/DataFormManager';
-  import type { DataFormField } from '../data-form-utilities/types';
+  import type { DataFormField } from '../data-form-utilities/FormFields';
 
   import { AddField } from './add-field';
   import FormFieldSource from './FormFieldSource.svelte';
