+++ mathesar_ui/src/systems/data-forms/elements/FormFieldSource.svelte
@@ -3,7 +3,7 @@
   import TableName from '@mathesar/components/TableName.svelte';
 
   import type { EditableDataFormManager } from '../data-form-utilities/DataFormManager';
-  import type { DataFormField } from '../data-form-utilities/types';
+  import type { DataFormField } from '../data-form-utilities/FormFields';
 
   export let dataFormManager: EditableDataFormManager;
   export let dataFormField: DataFormField;
